pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "synthetic-data-generation"
version = "0.1.0"
description = "A library for generating synthetic conversation data"
readme = "README.md"
requires-python = ">=3.7"
license = {file = "LICENSE"}
authors = [
    {name = "Channel Labs"}
]
dependencies = [
    "anthropic",
    "openai",
    "pyyaml",
]

[tool.setuptools.packages.find]
include = ["synthetic_conversation_generation*"]
namespaces = false
